+++ spec/remote_ruby/compiler_spec.rb
@@ -37,7 +37,7 @@ describe RemoteRuby::Compiler do
 
       it 'includes locals serialization' do
         client_locals.each_key do |name|
-          expect(compiled_code).to include("#{name} = __context__.unmarshal(:#{name}, ")
+          expect(compiled_code).to include("#{name} = __rr_context__.unmarshal(:#{name}, ")
         end
       end